refactor: isolate page related types
