+++ src/types/types.ts
@@ -1,19 +1,3 @@
-import { ReactElement, ReactNode } from 'react';
-import { NextPage } from 'next';
-
-// TODO: remove this pageMeta and NextPageWithLayout to use getStaticProps for layouts?
-export interface PageOptions {
-  title?: string;
-  useBackButton?: boolean;
-  rightNavMenu?: ReactNode;
-}
-
-// eslint-disable-next-line @typescript-eslint/ban-types
-export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
-  getLayout?: (page: ReactElement) => ReactNode;
-  pageOptions?: PageOptions;
-};
-
 export type UserType = 'user' | 'admin';
 
 export interface JwtToken {
